requirements.txt
Django==3.0.6
djangorestframework==3.11.0
Pillow==7.1.2
